0408 creater -> creator 오타 발견

+++ src/main/java/com/onrank/server/api/service/member/MemberService.java
@@ -57,9 +57,9 @@ public class MemberService {
     }
 
     /**
-     * 사용자가 특정 스터디에서 HOST 또는 CEREATER 역할을 가지고 있는지 확인
+     * 사용자가 특정 스터디에서 HOST 또는 CREATOR 역할을 가지고 있는지 확인
      */
-    public boolean isMemberCreaterOrHost(String username, Long studyId) {
+    public boolean isMemberCreatorOrHost(String username, Long studyId) {
         Student student = studentService.findByUsername(username)
                 .orElseThrow(() -> new IllegalArgumentException("Student not found"));
 
@@ -67,7 +67,7 @@ public class MemberService {
         return memberRepository.findByStudentStudentIdAndStudyStudyId(studentId, studyId)
                 .map(member ->
                         member.getMemberRole().equals(MemberRole.HOST) ||
-                        member.getMemberRole().equals(MemberRole.CREATER))
+                        member.getMemberRole().equals(MemberRole.CREATOR))
                 .orElse(false);
     }
 
@@ -108,9 +108,9 @@ public class MemberService {
         Member member = memberRepository.findByMemberIdAndStudyStudyId(memberId, studyId)
                 .orElseThrow(() -> new IllegalArgumentException("Member not in Study"));
 
-        // CREATER는 권한 변경 불가
-        if (member.getMemberRole() == MemberRole.CREATER) {
-            throw new IllegalStateException("CREATER는 권한을 수정할 수 없습니다.");
+        // CREATOR 는 권한 변경 불가
+        if (member.getMemberRole() == MemberRole.CREATOR) {
+            throw new IllegalStateException("CREATOR 는 권한을 수정할 수 없습니다.");
         }
         member.changeRole(MemberRole.valueOf(newRole));
     }
@@ -132,8 +132,9 @@ public class MemberService {
         }
 
         // HOST는 삭제 불가
-        if (targetMember.getMemberRole().equals(MemberRole.HOST)) {
-            throw new IllegalArgumentException("HOST는 삭제할 수 없습니다.");
+        if (targetMember.getMemberRole().equals(MemberRole.HOST)||
+                targetMember.getMemberRole().equals(MemberRole.CREATOR)) {
+            throw new IllegalArgumentException("HOST 와 CREATOR 는 삭제 불가");
         }
         memberRepository.delete(targetMember);
     }
